// src/app/page/master/kategori-gaji/kategori-gaji.interface.ts
import { EventEmitter } from '@angular/core';

export interface KategoriGaji {
    kode?: any;
    namaKategoryGaji?: string;
    kategoryGaji?: any;
    reportDisplay?: string;
    tanggalRevisi?: string;
    departemen?: string;
    noUrut?: string;
    kategoriGajiHead?: string;
    isAutoRevisiEvaluasi?: string;
     tglRevisiEvaluasiEfektif?: any;
    kodeExternal?: string;
    namaExternal?: string;
    statusEnabled?: boolean;
    noRec?: string;
    label?: any;
    value?: any;
    id?: any;
}